# backend/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8082
  description: Generated server url
paths:
...
  /api/v1/manage/clubs/{clubId}/matches/{matchId}/sides/{sideId}:
    put:
      tags:
      - match-side-controller
      operationId: updateSide
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: matchId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: sideId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMatchSideRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MatchSideDto"
  /api/v1/manage/clubs/{clubId}/matches/{matchId}/sides/{sideId}/players/{playerProfileId}:
    put:
      tags:
      - match-side-controller
      operationId: updatePlayerRole
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: matchId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: sideId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: playerProfileId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMatchSidePlayerRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MatchSideDto"
  /api/v1/manage/clubs/{clubId}/matches/{matchId}/sides/{sideId}/players/reorder:
    put:
      tags:
      - match-side-controller
      operationId: reorderPlayers
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: matchId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: sideId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReorderMatchSidePlayersRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MatchSideDto"
...
  /api/v1/manage/clubs/{clubId}/matches/{matchId}/sides:
    get:
      tags:
      - match-side-controller
      operationId: list_9
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: matchId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MatchSideDto"
    post:
      tags:
      - match-side-controller
      operationId: createSide
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: matchId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMatchSideRequest"
        required: true
      responses:
        "201":
          description: Match side created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MatchSideDto"
  /api/v1/manage/clubs/{clubId}/matches/{matchId}/sides/{sideId}/players:
    post:
      tags:
      - match-side-controller
      operationId: addPlayer
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: matchId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: sideId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddMatchSidePlayerRequest"
        required: true
      responses:
        "201":
          description: Player added to match side
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MatchSideDto"
  /api/v1/manage/clubs/{clubId}/matches/{matchId}/sides/{sideId}/players/{playerProfileId}/remove:
    post:
      tags:
      - match-side-controller
      operationId: removePlayer
      parameters:
      - name: clubId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: matchId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: sideId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: playerProfileId
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/MatchSideDto"
...
components:
  schemas:
...
    UpdateMatchSideRequest:
      type: object
      properties:
        captainPlayerId:
          type: string
          format: uuid
        wicketKeeperPlayerId:
          type: string
          format: uuid
        twelfthManPlayerId:
          type: string
          format: uuid
    MatchSideDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        matchId:
          type: string
          format: uuid
        teamId:
          type: string
          format: uuid
        captainPlayerId:
          type: string
          format: uuid
        wicketKeeperPlayerId:
          type: string
          format: uuid
        twelfthManPlayerId:
          type: string
          format: uuid
        players:
          type: array
          items:
            $ref: "#/components/schemas/MatchSidePlayerDto"
    MatchSidePlayerDto:
      type: object
      properties:
        playerProfileId:
          type: string
          format: uuid
        battingOrder:
          type: integer
          format: int32
        role:
          type: string
          enum:
          - BATSMAN
          - BOWLER
          - ALL_ROUNDER
    UpdateMatchSidePlayerRequest:
      type: object
      properties:
        role:
          type: string
          enum:
          - BATSMAN
          - BOWLER
          - ALL_ROUNDER
      required:
      - role
    ReorderMatchSidePlayersRequest:
      type: object
      properties:
        playerProfileIds:
          type: array
          items:
            type: string
            format: uuid
      required:
      - playerProfileIds
...
    CreateMatchSideRequest:
      type: object
      properties:
        teamId:
          type: string
          format: uuid
      required:
      - teamId
    AddMatchSidePlayerRequest:
      type: object
      properties:
        playerProfileId:
          type: string
          format: uuid
        role:
          type: string
          enum:
          - BATSMAN
          - BOWLER
          - ALL_ROUNDER
      required:
      - playerProfileId
      - role
...